+++ deployments/terraform/gcp/modules/network/variables.tf
@@ -1,0 +1,28 @@
+variable "project_id" {
+  type = string
+}
+
+variable "region" {
+  type = string
+}
+
+variable "name_prefix" {
+  type = string
+}
+
+variable "vpc_cidr" {
+  type = string
+}
+
+variable "pods_cidr" {
+  type = string
+}
+
+variable "services_cidr" {
+  type = string
+}
+
+variable "labels" {
+  type    = map(string)
+  default = {}
+}